Widen knowledge_sections.sort_key to bigint
Section sort keys are number * 10_000^2, so a section numbered 22+ exceeds
the 32-bit signed maximum. PostgreSQL enforces the INTEGER column and the
knowledge ingest failed on first deploy ('value out of int32 range');
SQLite in tests does not enforce it, which is why this only surfaced live.
The column is now BIGINT, matching what the sort formula has always
produced. Verified: alembic upgrade on pgvector, then a full ingest that
returned real DeKUT dean emails.

+++ backend/app/models/knowledge.py
@@ -132,7 +132,7 @@ class KnowledgeSection(Base, TimestampMixin):
     )
     # "1" for a top-level heading, "1.1" for a subsection.
     number: Mapped[str] = mapped_column(sa.String(12), nullable=False)
-    sort_key: Mapped[int] = mapped_column(sa.Integer, nullable=False, default=0)
+    sort_key: Mapped[int] = mapped_column(sa.BigInteger, nullable=False, default=0)
     title: Mapped[str] = mapped_column(sa.String(300), nullable=False, default="")
     body: Mapped[str] = mapped_column(sa.Text, nullable=False, default="")
     # Lowercased, punctuation-stripped title + body. Precomputed at write
